Major updates to the documentation
Major updates were made to the documentation, mostly regarding the
particle tracking component of PSOM:
- included the interp_bilinear subroutine
- added a warning about prescribing sinking velocity wrt the function
used to define vertical levels. This section refers to an added
appendix that includes the pdf of the document written about
prescribing sinking velocity to a particle; the issues, the error, the
solution and the final method used.
- Updated description of the matlab code that can be used to open the
particle-tracking binary files. the raw code of these MATLAB functions
were included as appendices.
- Updated the section on how to customize the particle-tracking code
- Added many entries in the glossary
- Added 3 appendices (see above)